+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.20.0
+Add `findApply`, `isNotNil`, `exists`, `unlessExists`, `unlessTruth`, `getOrReturn`, `alias`, `cascade`, `cascadeIn`, `cascadeKey`, `isMultiple`, `append`, `composeApply`, `comply`
+
+# 1.19.0
+Add ramda style `ifElse`, `where`, and `unless` to a new `logic` section
+
 # 1.18.1
 Update lensOf to use reduce instead of mapValues